Refactor echo service to use oslo_config
I need a bit more control over how I configure auth_token middleware
that can really only be done via oslo_config. Turn echo into a proper
service.
