+++ playbooks/roles/echo/templates/echo.conf
@@ -1,0 +1,20 @@
+[DEFAULT]
+debug = False
+
+
+[keystone_authtoken]
+auth_uri = http://{{ keystone_host | default(hostvars[groups["keystone"][0]]["ansible_default_ipv4"]["address"]) }}:35357/
+memcached_servers = 127.0.0.1:11211
+token_cache_time = 300
+revocation_cache_time = 60
+memcache_security_strategy = ENCRYPT
+memcache_secret_key = 5bdc5pilx0o7re34z3yc2yt4x
+include_service_catalog =  True
+
+auth_plugin = password
+auth_url = http://{{ keystone_host | default(hostvars[groups["keystone"][0]]["ansible_default_ipv4"]["address"]) }}:35357/
+username = admin
+user_domain_id = default
+password = password
+project_name = admin
+project_domain_id = default
